Implement ShoppingLib into WitchSlayer-Bot, Flint-Bot and Magmus-Bot
ShoppingLib:
-Bug-Fix: rare purchase error (Firebrand into Dawnbringer syndrom)
-comments

Example bots removed
