--- bots/advancedShoppingLib/exampleBots/ASflint.bot
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<bot
-	name="AdvancedShopping-FlintBot"
-	description="bot_description_flint_beastwood"
-	heroname="Hero_FlintBeastwood"
-	rootlua="ASflint_main.lua"
-	defaultplayername="FlintBot"
-	precachepath="/heroes/flint_beastwood/"
-/>
