--- bots/advancedShoppingLib/exampleBots/ASwitchslayer.bot
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<bot
-	name="AdvancedShopping-WitchSlayerBot"
-	description="bot_description_witch_slayer"
-	heroname="Hero_WitchSlayer"
-	rootlua="ASwitchslayer_main.lua"
-	defaultplayername="WitchSlayerBot"
-	precachepath="/heroes/witch_slayer/"
-/>
